Tested against matlab script
Important change to file handling. If "run" appears in the file name for prep0 (the first file name in the directory) the script will look for multiple runs and analyze all the data, otherwise it will toss all the file names with "run" in the name and analyze only the set without "run". This serves as a switch if you want to manually remove some files from consideration. The only catch is that it is all or nothing.

I haven't yet figured out how to best do the motion detection handling. Currently the strategy is to inspect the results of the initial run and then run the script again modifying the "manual motion removal block."

I compared sub-SCD01 and sub-SCD007 against the old matlab script and they compare favorably. A slight 1 part in a thousand T2 difference, but that is likely a result of the fitting or a rounding error. I am satisfied that those to subjects give identical results.

In the process I discovered an error in the matlab script that is now flagged. It means that the SCD analysis was not quite right. Since I am satisfied that this script is now ready for prime time I will use it to reanalyze the data.